Extract shared Jackson abstractions into jackson module
Move the format-agnostic Jackson pieces out of the json module into a
new module.

This lets json and the upcoming yaml module depend on shared code as
siblings instead of yaml reaching into json's internals.
