+++ api/pom.xml
@@ -28,6 +28,11 @@
 
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>jackson</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>jaxrs</artifactId>
         </dependency>
