+++ jackson/src/main/java/io/airlift/jackson/JacksonSubTypeBinder.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.airlift.json;
+package io.airlift.jackson;
 
 import com.google.inject.Binder;
 import com.google.inject.multibindings.Multibinder;
