+++ jackson/src/main/java/io/airlift/jackson/RecordAutoDetectModule.java
@@ -1,4 +1,4 @@
-package io.airlift.json;
+package io.airlift.jackson;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
