+++ json/src/main/java/io/airlift/json/ObjectMapperProvider.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.core.JsonFactoryBuilder;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
+import io.airlift.jackson.BaseJacksonProvider;
 
 import static java.util.Objects.requireNonNull;
