+++ mcp/src/main/java/io/airlift/mcp/McpModule.java
@@ -15,8 +15,8 @@ import com.google.inject.multibindings.Multibinder;
 import com.google.inject.multibindings.OptionalBinder;
 import com.google.inject.name.Names;
 import com.google.inject.util.Types;
-import io.airlift.json.JacksonSubType;
-import io.airlift.json.JacksonSubTypeBinder;
+import io.airlift.jackson.JacksonSubType;
+import io.airlift.jackson.JacksonSubTypeBinder;
 import io.airlift.mcp.handler.CompletionEntry;
 import io.airlift.mcp.handler.PromptEntry;
 import io.airlift.mcp.handler.ResourceEntry;
@@ -72,7 +72,7 @@ import static com.google.inject.Scopes.SINGLETON;
 import static com.google.inject.multibindings.Multibinder.newSetBinder;
 import static com.google.inject.multibindings.OptionalBinder.newOptionalBinder;
 import static io.airlift.configuration.ConfigBinder.configBinder;
-import static io.airlift.json.JacksonSubTypeBinder.jacksonSubTypeBinder;
+import static io.airlift.jackson.JacksonSubTypeBinder.jacksonSubTypeBinder;
 import static io.airlift.json.JsonBinder.jsonBinder;
 import static io.airlift.mcp.reflection.ReflectionHelper.forAllInClass;
 import static io.airlift.mcp.reflection.SkillsHelper.resourceFromSkill;
